lambda/requirements.txt
icrawler
boto3
pillow
pinecone
python-dotenv
